refactor(cli): move tracing config under reporting
