+++ crates/tsz-cli/src/lib.rs
@@ -14,12 +14,12 @@ pub mod project;
 pub mod project_refs;
 pub mod reporter;
 pub mod reporting;
-pub mod tracing_config;
 pub mod watch;
 pub use commands::build;
 pub use commands::help;
 pub use project::fs;
 pub use reporting::trace;
+pub use reporting::tracing_config;
 
 #[cfg(test)]
 #[path = "../tests/args_tests.rs"]
