+++ crates/tsz-cli/src/reporting/mod.rs
@@ -1,1 +1,2 @@
 pub mod trace;
+pub mod tracing_config;
